- add GetMapping, - improve validating methods, - improve updating methods, - add AppConfiguration class, - improve ResourceMapper class.

+++ src/main/java/pl/gispartner/ResourceManagerApp/service/ResourceService.java
@@ -4,14 +4,12 @@ import com.fasterxml.jackson.databind.JsonNode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import pl.gispartner.ResourceManagerApp.model.ResourceDto;
-import pl.gispartner.ResourceManagerApp.model.ResourceEntity;
-import pl.gispartner.ResourceManagerApp.model.ResourceMapper;
-import pl.gispartner.ResourceManagerApp.model.UserEntity;
+import pl.gispartner.ResourceManagerApp.model.*;
 import pl.gispartner.ResourceManagerApp.persistance.ResourceRepository;
 import pl.gispartner.ResourceManagerApp.persistance.UserRepository;
 
 import java.util.Date;
+import java.util.List;
 import java.util.Objects;
 
 import static pl.gispartner.ResourceManagerApp.model.UserType.SUPER_USER;
@@ -22,40 +20,66 @@ public class ResourceService {
 
     private final ResourceRepository resourceRepository;
     private final UserRepository userRepository;
+    private final ResourceMapper resourceMapper;
+
+    public ResourceDto getResource(Long resourceId) {
+        return resourceMapper.mapToDto(resourceRepository.getReferenceById(resourceId));
+    }
+
+    public List<ResourceDto> getAllUserResources(Long userId) {
+        UserEntity userEntity = userRepository.getReferenceById(userId);
+        return userEntity.getResources().stream()
+                .map(resourceMapper::mapToDto)
+                .toList();
+    }
 
     public Long saveResource(ResourceDto resourceDto) {
-        ResourceEntity resourceEntity = ResourceMapper.mapToEntity(resourceDto);
+        ResourceEntity resourceEntity = resourceMapper.mapToEntity(resourceDto);
         UserEntity userEntity = userRepository.getReferenceById(resourceDto.getOwnerId());
 
         resourceEntity.setUser(userEntity);
-        resourceEntity.setJsonData(ResourceMapper.saveToString(resourceDto.getJsonData()));
+        resourceEntity.setJsonData(resourceMapper.saveToString(resourceDto.getJsonData()));
 
         ResourceEntity created = resourceRepository.save(resourceEntity);
         return created.getId();
     }
 
-    public void deleteResource(Long resourceId) {
+    public String deleteResource(Long resourceId, Long userId) {
+        if (!isUserValid(userId, resourceId)) {
+            return "This operation cannot be performed";
+        }
         resourceRepository.deleteById(resourceId);
+        return "Changes have been successfully saved";
     }
 
     @Transactional
-    public void updateResourceName(Long resourceId, String newResourceName) {
-        resourceRepository.updateResourceName(resourceId, newResourceName);
-        ResourceEntity resourceEntity = resourceRepository.getReferenceById(resourceId);
+    public String updateResourceName(Long resourceId, String newResourceName, Long userId) {
+        if (!isUserValid(userId, resourceId)) {
+            return "This operation cannot be performed";
+        }
+        ResourceEntity resourceEntity = resourceRepository.findById(resourceId)
+                .orElseThrow(() -> new RuntimeException("Resource not found"));
+        resourceEntity.setName(newResourceName);
         resourceEntity.setModifiedDate(new Date());
+        return "Changes have been successfully saved";
     }
 
     @Transactional
-    public void updateJsonData(Long resourceId, JsonNode newJsonData) {
-        resourceRepository.updateJsonData(resourceId, newJsonData);
-        ResourceEntity resourceEntity = resourceRepository.getReferenceById(resourceId);
+    public String updateJsonData(Long resourceId, JsonNode newJsonData, Long userId) {
+        if (!isUserValid(userId, resourceId)) {
+            return "This operation cannot be performed";
+        }
+        ResourceEntity resourceEntity = resourceRepository.findById(resourceId)
+                .orElseThrow(() -> new RuntimeException("Resource not found"));
+        resourceEntity.setJsonData(resourceMapper.saveToString(newJsonData));
         resourceEntity.setModifiedDate(new Date());
+        return "Changes have been successfully saved";
     }
 
-    public boolean validateResourceOwner(Long userId, Long resourceId) {
+    public boolean isUserValid(Long userId, Long resourceId) {
         ResourceEntity resourceEntity = resourceRepository.getReferenceById(resourceId);
         UserEntity userEntity = userRepository.getReferenceById(userId);
-        return Objects.equals(userId, resourceEntity.getUser().getId()) || userEntity.getUserType().equals(SUPER_USER);
+        return Objects.equals(userId, resourceEntity.getUser().getId()) || Objects.equals(userEntity.getUserType(), SUPER_USER);
     }
 
 }
